requirements.txt
asgiref==3.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-redis==5.2.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
idna==3.4
psycopg2-binary==2.9.6
redis==4.5.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
django-celery-results==2.5.1
django-celery-beat==2.5.0
pytest-django
httpx==0.24.1
Pillow==10.1.0
django-filter==22.1
pre-commit==3.5.0
djangorestframework-simplejwt==5.3.0
